split ZMQConf into ZMQConf and child class PublisherSubscriberConf, and rename attributes
Strictly speaking the conf value `packer` is not an ZMQ config value,
but only a config value for Publishers and Subscribers.
As the broker will use the ZMQConf as well it is split into two,
and the PublisherSubscriberConf will extend ZMQConf.
Additionally rename producer_port->publisher_port,
consumer_port->subscriber_port and address->interface,
as this follows the ZMQ terminology.